requirements.txt
Django==1.10.5
django-debug-toolbar==1.6
django-registration-redux==1.4
django-tokenapi==0.2.5
psycopg2==2.6.2
gunicorn==19.6.0
